Create base class for VideoFrameProvider
